pod/pdm attributes moved into definitions.py
